backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http